- separate component into folders
