+++ app/javascript/components/common/NotificationToast.jsx
