+++ app/javascript/components/dashboard/Dashboard.jsx
@@ -1,0 +1,13 @@
+import React from "react"
+import PropTypes from "prop-types"
+class Dashboard extends React.Component {
+  render() {
+    return (
+      <React.Fragment>
+        <h1 className="dab">HI</h1>
+      </React.Fragment>
+    );
+  }
+}
+
+export default Dashboard
