+++ app/javascript/components/external/ExternalFooter.jsx
