+++ app/javascript/components/external/ExternalHeader.jsx
